backend/requirements.txt
alembic==1.14.0
APScheduler==3.11.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
icmplib==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
pyasn1==0.6.1
PyJWT==2.10.1
PyMySQL==1.1.1
pysnmp==7.1.15
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Werkzeug==3.1.3
wsproto==1.2.0
